examples/social/shape-irc/pyproject.toml
[project]
name = "shapesinc-python-irc"
version = "1.0.0"
description = "IRC integration using the Shapes API."
authors = [
    { name = "Shapes Inc.", email = "hi@shapes.inc" }
]
keywords = []
license = { text = "MIT" }
readme = "README.md"
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
]
requires-python = ">=3.11"
dependencies = [
    "irc>=20.0.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.8.0"
]

[project.urls]
Homepage = "https://shapes.inc"
Repository = "https://github.com/shapesinc/api.git"

[project.scripts]
shape-irc = "main:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
